requirements.txt
python-telegram-bot
expiringdict
google-api-python-client
google-auth-httplib2
google-auth-oauthlib